# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Galaktika galaxy game",
        "title": "Galaktika API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/divisions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "divisions"
                ],
                "summary": "List all divisions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/galaxy.Division"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "divisions"
                ],
                "summary": "Create a division",
                "parameters": [
                    {
                        "description": "Division data",
                        "name": "division",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.Division"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/galaxy.Division"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/divisions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "divisions"
                ],
                "summary": "Get a division by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Division ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.Division"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "divisions"
                ],
                "summary": "Update a division",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Division ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Division data",
                        "name": "division",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.Division"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.Division"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "divisions"
                ],
                "summary": "Delete a division",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Division ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "galaxy.Division": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "resources_amount": {
                    "type": "integer"
                },
                "tech_attack": {
                    "type": "integer"
                },
                "tech_cargo": {
                    "type": "integer"
                },
                "tech_defense": {
                    "type": "integer"
                },
                "tech_engines": {
                    "type": "integer"
                }
            }
        },
...
    }
}